requirements.txt
Package                             Version
----------------------------------- ---------
aniso8601                           9.0.1
appdirs                             1.4.3
asgiref                             3.7.2
black                               20.8b1
CacheControl                        0.12.6
certifi                             2023.7.22
cffi                                1.15.1
cfgv                                3.3.0
chardet                             3.0.4
charset-normalizer                  3.2.0
click                               7.1.2
colorama                            0.4.3
contextlib2                         0.6.0
coreapi                             2.3.3
coreschema                          0.0.4
cryptography                        3.2.1
defusedxml                          0.7.0rc1
distlib                             0.3.7
distro                              1.4.0
Django                              4.1.1
django-3-jet                        1.0.8
django-braces                       1.14.0
django-cors-headers                 3.11.0
django-elasticsearch-dsl            7.2.2
django-environ                      0.11.2
django-filter                       23.3
django-graphql-jwt                  0.3.4
django-jazzmin                      2.6.0
django-oauth-toolkit                1.3.3
django-prometheus                   2.2.0
django-rest-framework-social-oauth2 1.1.0
django-widget-tweaks                1.5.0
djangorestframework                 3.13.1
djangorestframework-simplejwt       5.0.0
djongo                              1.3.6
dnspython                           2.4.2
drf-yasg                            1.20.0
elastic-transport                   8.4.0
elasticsearch                       7.13.4
elasticsearch-dsl                   7.4.0
filelock                            3.12.4
graphene                            3.0
graphene-django                     3.0.0b7
graphene-file-upload                1.3.0
graphql-core                        3.1.7
graphql-relay                       3.1.0
gunicorn                            20.0.4
html5lib                            1.0.1
identify                            2.2.10
idna                                3.4
inflection                          0.3.1
ipaddr                              2.2.0
isort                               5.9.1
itypes                              1.1.0
Jinja2                              2.11.1
lockfile                            0.12.2
MarkupSafe                          1.1.1
msgpack                             0.6.2
mypy-extensions                     0.4.3
nodeenv                             1.6.0
oauthlib                            3.1.0
openapi-codec                       1.3.2
packaging                           20.1
pathspec                            0.8.1
pep517                              0.8.2
Pillow                              9.3.0
pip                                 23.2.1
platformdirs                        3.10.0
pre-commit                          2.13.0
progress                            1.5
prometheus-client                   0.14.1
promise                             2.3
pycparser                           2.20
Pygments                            2.7.1
PyJWT                               2.3.0
pymongo                             3.12.3
pyparsing                           2.4.6
python-dateutil                     2.8.1
python3-openid                      3.2.0
pytoml                              0.1.21
pytz                                2018.9
PyYAML                              6.0.1
regex                               2023.8.8
requests                            2.31.0
requests-oauthlib                   1.3.0
retrying                            1.3.3
ruamel.yaml                         0.16.10
ruamel.yaml.clib                    0.2.0
Rx                                  1.6.1
setuptools                          68.2.2
simplejson                          3.17.0
singledispatch                      3.7.0
six                                 1.14.0
social-auth-app-django              5.0.0
social-auth-core                    4.1.0
sqlparse                            0.2.4
text-unidecode                      1.3
toml                                0.10.2
typed-ast                           1.5.5
typing-extensions                   3.7.4.3
tzdata                              2023.3
uritemplate                         3.0.1
urllib3                             2.0.5
virtualenv                          20.24.5
webencodings                        0.5.1
whitenoise                          5.0.1